requirements.txt
numpy
matplotlib
pyrtlsdr
pyqtgraph
